+++ R/NAMESPACE.R
@@ -14,9 +14,12 @@
 #' @importFrom colourpicker colourInput 
 #' @importFrom DataTools checkAnyNonNumericColumns downloadModelUI downloadModelServer 
 #'  importDataUI importDataServer importOptions renameExistingNames tryCatchWithWarningsAndErrors
-#' @importFrom dplyr arrange bind_rows distinct slice
-#' @importFrom ggplot2 aes element_line element_text ggplot geom_line geom_point geom_ribbon 
-#' labs scale_x_continuous theme ggtitle scale_y_continuous geom_vline coord_cartesian sec_axis
+#' @importFrom dplyr arrange bind_cols bind_rows cur_group_id distinct do group_by mutate n select
+#'  slice ungroup
+#' @importFrom ggplot2 aes coord_cartesian element_line element_text 
+#'  geom_line geom_point geom_ribbon geom_vline ggplot ggplot_build ggtitle labs    
+#'  scale_colour_manual scale_fill_manual scale_shape_manual scale_size_manual 
+#'  scale_x_continuous scale_y_continuous sec_axis theme
 #' @importFrom htmltools save_html
 #' @importFrom jsonlite toJSON
 #' @importFrom magrittr %>%
@@ -24,11 +27,13 @@
 #' @importFrom parallel detectCores
 #' @importFrom rlang .data
 #' @importFrom rstan sampling extract
+#' @importFrom shinycssloaders withSpinner
 #' @importFrom shinyjs alert
-#' @importFrom shinyTools formatPointsOfGGplot formatRangesOfGGplot formatTitlesOfGGplot
+#' @importFrom shinyTools dataExportButton dataExportServer formatPointsOfGGplot
+#'  formatRangesOfGGplot formatTitlesOfGGplot plotExportButton plotExportServer
 #'  plotPointsServer plotPointsUI plotRangesServer plotRangesUI plotTitlesServer plotTitlesUI
 #' @importFrom shinyWidgets pickerInput updatePickerInput
-#' @importFrom stats approx dnorm lm median quantile sd
+#' @importFrom stats approx dnorm lm median na.omit quantile sd
 #' @importFrom utils write.csv write.table combn
 #' @importFrom yaml read_yaml
 #' @references